+++ src/components/Clock.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Route } from 'react-router-dom';
 
 export default function Clock() {
   return <div>Clock</div>;
